Minor refactor of Extraction* classes
* Moved extraction parameters to separate files
* Added some error checks/messages
