+++ Source/ParticleInterpolator/SurfaceExtraction.hpp
@@ -9,6 +9,9 @@
 // AMReX includes
 #include <AMReX_GpuContainers.H>
 
+// Parameters
+#include "SurfaceExtractionParameters.hpp"
+
 // Other includes
 #include "DimensionDefinitions.hpp"
 #include "FilesystemTools.hpp"
@@ -25,35 +28,6 @@
 #include <functional>
 #include <utility>
 #include <vector>
-
-template <int num_components> class ParticleInterpolator;
-
-struct surface_extraction_params_t
-{
-    int num_surfaces{}; //!< number of surfaces over which to extraction
-    amrex::Gpu::ManagedVector<double>
-        surface_param_values; //!< the values of the
-                              //!< parameter that gives the required
-                              //!< surfaces with SurfaceGeom geometry (e.g.
-                              //!< radii for spherical shells)
-    int num_points_u{};       //!< the number of points for the first parameter
-                              //!< that parameterises each surface
-    int num_points_v{};       //!< the number of points for the second parameter
-                              //!< that parameterises each surfaces
-    amrex::Gpu::ManagedVector<int>
-        extraction_levels;   //!< the level on which to do the
-                             //!< extraction for each surface
-    bool write_extraction{}; //!< whether or not to write the extracted data
-
-    std::string data_path, integral_file_prefix;
-    std::string extraction_path, extraction_file_prefix;
-
-    int min_extraction_level() const
-    {
-        return *(std::min_element(extraction_levels.begin(),
-                                  extraction_levels.end()));
-    }
-};
 
 //! This class extracts grid variables on 2 dimensional surfaces each
 //! parameterised by u and v with different surfaces given by level sets of
